+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added checks for pin allocation, trying to take output pins 0-2 while using
   the SPI interface will now result in panic.
-- Added the ability to 
+- Added I2C traits.
+
+### Changed
+- Changed the FTDI MPSSE initialization to occur once globally for the device
+  instead of inside the SPI device trait.
 
 ## [0.4.0] - 2021-03-05
 ### Added